#5 la balle bouge trés aleatoirement, les colissions sont faites uniquement par rapport au bord de la fenetre

+++ src/main/java/entite/Racket.java
@@ -63,7 +63,7 @@ public class Racket extends Entite{
         this.c.setY(pY);
     }
 
-    public int mX(){
+    public double mX(){
         return this.c.getX();
     }
