+++ src/main/java/geometry/Coordonnee.java
@@ -1,31 +1,31 @@
 package geometry;
 
 public class Coordonnee {
-    int x;
-    int y;
+    double x;
+    double y;
 
-    public Coordonnee(int x, int y) {
+    public Coordonnee(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x){
+    public void setX(double x){
         this.x=x;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setY(int y){
+    public void setY(double y){
         this.y=y;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setCoordonnee(int x, int y){
+    public void setCoordonnee(double x, double y){
         this.x=x;
         this.y=y;
     }
